+++ ApiPaymets/RequisitonsModels/Payloads/PaymentPayload.cs
@@ -1,5 +1,5 @@
 ﻿namespace ApiPayments.RequisitionsModels.Payloads
 {
-    public sealed record PaymentPayloadModel(Guid correlationId, float amount);
-    public record PaymentPayloadRequestModel(Guid correlationId, float amount, DateTime requestedAt);
+    public sealed record PaymentPayloadModel(Guid correlationId, decimal amount);
+    public record PaymentPayloadRequestModel(Guid correlationId, decimal amount, DateTime requestedAt);
 }
